Start thinking about new interface
