+++ lib/bugsnag/integrations/delayed_job.rb
