+++ lib/bugsnag/integrations/rack.rb
@@ -6,7 +6,7 @@ module Bugsnag
       # Configure bugsnag rack defaults
       Bugsnag.configure do |config|
         # Try to set the release_stage automatically if it hasn't already been set
-        config.release_stage ||= release_stage
+        config.release_stage ||= ENV["RACK_ENV"] if ENV["RACK_ENV"]
 
         # Try to set the project_root if it hasn't already been set, or show a warning if we can't
         unless config.project_root && !config.project_root.to_s.empty?
@@ -27,7 +27,7 @@ module Bugsnag
 
     def call(env)
       # Set the request data for bugsnag middleware to use
-      Bugsnag.set_request_data(:rack_env, env)
+      Bugsnag.configuration.set_request_data(:rack_env, env)
 
       begin
         response = @app.call(env)
@@ -47,13 +47,7 @@ module Bugsnag
       response
     ensure
       # Clear per-request data after processing the each request
-      Bugsnag.clear_request_data
-    end
-
-    private
-
-    def release_stage
-      ENV["BUGSNAG_RELEASE_STAGE"] || ENV["RACK_ENV"]
+      Bugsnag.configuration.clear_request_data
     end
   end
 end
