+++ lib/bugsnag/integrations/rails/active_record_rescue.rb
