+++ lib/bugsnag/integrations/resque.rb
