+++ lib/bugsnag/integrations/sidekiq.rb
@@ -11,7 +11,7 @@ module Bugsnag
     def call(worker, msg, queue)
       begin
         # store msg/queue in thread local state to be read by Bugsnag::Middleware::Sidekiq
-        Bugsnag.set_request_data :sidekiq, { :msg => msg, :queue => queue }
+        Bugsnag.configuration.set_request_data :sidekiq, { :msg => msg, :queue => queue }
 
         yield
       rescue Exception => ex
@@ -19,7 +19,7 @@ module Bugsnag
         Bugsnag.auto_notify(ex)
         raise
       ensure
-        Bugsnag.clear_request_data
+        Bugsnag.configuration.clear_request_data
       end
     end
   end
