+++ lib/bugsnag/meta_data.rb
@@ -3,5 +3,6 @@ module Bugsnag
     attr_accessor :bugsnag_meta_data
     attr_accessor :bugsnag_user_id
     attr_accessor :bugsnag_context
+    attr_accessor :bugsnag_grouping_hash
   end
 end
